\name{psm}
\alias{psm}
\docType{data}

\title{ 15N example data }

\description{
  An example data for 15N metabolic labelling, distributed as an
  \code{MSnSet} to illustrate 15N QC. }

\usage{data(psm)}

\references{
  See \code{\linkS4class{MSnSet}} and the \code{MSnbase-demo} vignette.   
}

\examples{
data(psm)
psm
}

\keyword{datasets}
