\name{psm}
\alias{psm}
\alias{n15psm}
\docType{data}

\title{ 15N example data }

\description{
  An example data for 15N metabolic labelling, distributed as an
  \code{MSnSet} to illustrate 15N QC. }

\usage{data(n15psm)}

\references{
  See \code{MSnSet} and the \code{MSnbase-demo} vignette in the
  \code{MSnbase} package.
}

\examples{
library("MSnbase")
data(n15psm)
psm
}

\keyword{datasets}
